jira-ai-cli/pyproject.toml
[project]
name = "jira-ai-cli"
version = "0.1.0"
description = "LLM-Assisted Jira CLI"
dependencies = [
    "click",
    "requests",
    "jira",
    "PyYAML",
]

[project.scripts]
jira-ai = "jira_ai_cli.cli:cli"

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
